Refactored, separated classes (Factorial, Exponentiation)

+++ AlgoChallange/Combinatorial/Binomial Coefficient/BinomialCoefficient.cs
@@ -2,6 +2,6 @@
 {
     public class BinomialCoefficient
     {
-        public static ulong Calculate(ulong n, ulong k, IBinomialCoefficientCalculator calculator) => calculator.Calculate(n, k);
+        public static ulong Calculate(uint n, uint k, IBinomialCoefficientCalculator calculator) => calculator.Calculate(n, k);
     }
 }
