+++ AlgoChallange/Combinatorial/Binomial Coefficient/IBinomialCoefficientCalculator.cs
@@ -2,6 +2,6 @@
 {
     public interface IBinomialCoefficientCalculator
     {
-        ulong Calculate(ulong n, ulong k);
+        ulong Calculate(uint n, uint k);
     }
 }
